+++ test_runner/src/main/kotlin/ftl/args/IosArgs.kt
@@ -52,6 +52,7 @@ class IosArgs(
     override val shardTime = cli?.shardTime ?: flank.shardTime
     override val repeatTests = cli?.repeatTests ?: flank.repeatTests
     override val smartFlankGcsPath = flank.smartFlankGcsPath
+    override val smartFlankUploadEnabled = cli?.smartFlankUploadEnabled ?: flank.smartFlankUploadEnabled
     override val testTargetsAlwaysRun = cli?.testTargetsAlwaysRun ?: flank.testTargetsAlwaysRun
     override val filesToDownload = cli?.filesToDownload ?: flank.filesToDownload
     override val disableSharding = cli?.disableSharding ?: flank.disableSharding
@@ -124,6 +125,7 @@ ${devicesToString(devices)}
       shard-time: $shardTime
       repeat-tests: $repeatTests
       smart-flank-gcs-path: $smartFlankGcsPath
+      smart-flank-upload-enabled: $smartFlankUploadEnabled
       test-targets-always-run:
 ${listToString(testTargetsAlwaysRun)}
       files-to-download: