+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/ios/IosRunCommand.kt
@@ -220,4 +220,10 @@ class IosRunCommand : Runnable {
         description = ["Saves test result to this local folder. Deleted before each run."]
     )
     var localResultsDir: String? = null
+
+    @Option(
+        names = ["--smart-flank-upload-enabled"],
+        description = ["Whether the JUnit XML results should be uploaded to the Smart Flank cloud storage."]
+    )
+    var smartFlankUploadEnabled: Boolean? = null
 }