+++ test_runner/src/test/kotlin/ftl/cli/firebase/test/android/AndroidRunCommandTest.kt
@@ -88,6 +88,7 @@ class AndroidRunCommandTest {
         assertThat(cmd.flakyTestAttempts).isNull()
         assertThat(cmd.disableSharding).isNull()
         assertThat(cmd.localResultDir).isNull()
+        assertThat(cmd.smartFlankUploadEnabled).isNull()
     }
 
     @Test
@@ -319,5 +320,13 @@ class AndroidRunCommandTest {
         CommandLine(cmd).parse("--local-result-dir=a")
 
         assertThat(cmd.localResultDir).isEqualTo("a")
+    }
+
+    @Test
+    fun `smartFlankUploadEnabled parse`() {
+        val cmd = AndroidRunCommand()
+        CommandLine(cmd).parse("--smart-flank-upload-enabled=false")
+
+        assertThat(cmd.smartFlankUploadEnabled).isEqualTo(false)
     }
 }